refactor(perf): split /ai-architecture into server shell + motion island
Fourth top-of-funnel page migrated to server shell pattern.
Server handles metadata + JSON-LD. Client island holds motion + state.
